pyproject.toml
[project]
name = "seo_article_agent"
version = "0.2.0"
description = "MCP対応AIエージェントがSEO最適化された人間らしい記事を作成します。"
requires-python = ">=3.11"
readme = "README.md"
authors = [{name="Your Name", email="you@example.com"}]

dependencies = [
    # コア
    "openai>=1.75.0",
    "openai-agents>=0.0.11",
    "tiktoken>=0.5.2",

    # クローリング / パース
    "beautifulsoup4>=4.12.3",
    "readability-lxml>=0.8.1",
    "aiohttp>=3.9.5",
    "lxml[html-clean]>=5.3.2",
    "python-slugify>=8.0.1",

    # MCP (stdio + 高速SSEサーバー)
    "fastmcp>=0.4.3",   # オプション、開発SSEデモ用のみ使用
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[project.scripts]
seo-agent = "seo_agent.cli:main"
